Refactor checkpointing Effect services
