+++ apps/server/src/checkpointing/CheckpointDiffQuery.ts
@@ -1,23 +1,55 @@
+/**
+ * CheckpointDiffQuery - Query interface for computed checkpoint diffs.
+ *
+ * Provides read-only diff operations across checkpoint snapshots used by
+ * orchestration APIs.
+ *
+ * @module CheckpointDiffQuery
+ */
 import {
   type CheckpointRef,
   OrchestrationGetTurnDiffResult,
-  type ThreadId,
+  type OrchestrationGetFullThreadDiffInput,
   type OrchestrationGetFullThreadDiffResult,
+  type OrchestrationGetTurnDiffInput,
   type OrchestrationGetTurnDiffResult as OrchestrationGetTurnDiffResultType,
+  type ThreadId,
 } from "@t3tools/contracts";
+import * as Context from "effect/Context";
 import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 import * as Option from "effect/Option";
 import * as Schema from "effect/Schema";
 
-import { ProjectionSnapshotQuery } from "../../orchestration/Services/ProjectionSnapshotQuery.ts";
-import { CheckpointInvariantError, CheckpointUnavailableError } from "../Errors.ts";
-import { checkpointRefForThreadTurn } from "../Utils.ts";
-import { CheckpointStore } from "../Services/CheckpointStore.ts";
-import {
+import * as ProjectionSnapshotQuery from "../orchestration/Services/ProjectionSnapshotQuery.ts";
+import { CheckpointInvariantError, CheckpointUnavailableError } from "./Errors.ts";
+import type { CheckpointServiceError } from "./Errors.ts";
+import { checkpointRefForThreadTurn } from "./Utils.ts";
+import * as CheckpointStore from "./CheckpointStore.ts";
+
+/** Service tag for checkpoint diff queries. */
+export class CheckpointDiffQuery extends Context.Service<
   CheckpointDiffQuery,
-  type CheckpointDiffQueryShape,
-} from "../Services/CheckpointDiffQuery.ts";
+  {
+    /**
+     * Read the patch diff for a single turn checkpoint transition.
+     *
+     * Verifies checkpoint availability in both projection state and filesystem.
+     */
+    readonly getTurnDiff: (
+      input: OrchestrationGetTurnDiffInput,
+    ) => Effect.Effect<OrchestrationGetTurnDiffResultType, CheckpointServiceError>;
+
+    /**
+     * Read the full patch diff across a thread range of checkpoints.
+     *
+     * Uses turn-diff semantics with `fromTurnCount = 0`.
+     */
+    readonly getFullThreadDiff: (
+      input: OrchestrationGetFullThreadDiffInput,
+    ) => Effect.Effect<OrchestrationGetFullThreadDiffResult, CheckpointServiceError>;
+  }
+>()("t3/checkpointing/CheckpointDiffQuery") {}
 
 const isTurnDiffResult = Schema.is(OrchestrationGetTurnDiffResult);
 
@@ -37,11 +69,11 @@ function buildTurnDiffResult(
   };
 }
 
-const make = Effect.gen(function* () {
-  const projectionSnapshotQuery = yield* ProjectionSnapshotQuery;
-  const checkpointStore = yield* CheckpointStore;
+export const make = Effect.gen(function* () {
+  const projectionSnapshotQuery = yield* ProjectionSnapshotQuery.ProjectionSnapshotQuery;
+  const checkpointStore = yield* CheckpointStore.CheckpointStore;
 
-  const getTurnDiff: CheckpointDiffQueryShape["getTurnDiff"] = Effect.fn("getTurnDiff")(
+  const getTurnDiff: CheckpointDiffQuery["Service"]["getTurnDiff"] = Effect.fn("getTurnDiff")(
     function* (input) {
       const operation = "CheckpointDiffQuery.getTurnDiff";
       const ignoreWhitespace = input.ignoreWhitespace ?? true;
@@ -145,7 +177,7 @@ const make = Effect.gen(function* () {
     },
   );
 
-  const getFullThreadDiff: CheckpointDiffQueryShape["getFullThreadDiff"] = Effect.fn(
+  const getFullThreadDiff: CheckpointDiffQuery["Service"]["getFullThreadDiff"] = Effect.fn(
     "CheckpointDiffQuery.getFullThreadDiff",
   )(function* (input) {
     const operation = "CheckpointDiffQuery.getFullThreadDiff";
@@ -239,10 +271,10 @@ const make = Effect.gen(function* () {
     return turnDiff satisfies OrchestrationGetFullThreadDiffResult;
   });
 
-  return {
+  return CheckpointDiffQuery.of({
     getTurnDiff,
     getFullThreadDiff,
-  } satisfies CheckpointDiffQueryShape;
+  });
 });
 
-export const CheckpointDiffQueryLive = Layer.effect(CheckpointDiffQuery, make);
+export const layer = Layer.effect(CheckpointDiffQuery, make);
