+++ oxlint-plugin-t3code/rules/no-manual-effect-runtime-in-tests.ts
@@ -25,7 +25,6 @@ const LEGACY_BASELINE = new Map<string, number>([
   ["apps/mobile/src/features/agent-awareness/liveActivityPreferences.test.ts", 1],
   ["apps/mobile/src/features/agent-awareness/remoteRegistration.test.ts", 2],
   ["apps/mobile/src/state/use-remote-environment-registry.test.ts", 2],
-  ["apps/server/src/checkpointing/Layers/CheckpointDiffQuery.test.ts", 5],
   ["apps/server/src/orchestration/commandInvariants.test.ts", 6],
   ["apps/server/src/orchestration/Layers/CheckpointReactor.test.ts", 42],
   ["apps/server/src/orchestration/Layers/OrchestrationEngine.test.ts", 5],
